video: tegra: nvmap: Add NVMAP_ELEM_SIZE_U64 macro
Support reserve and cache list operations for handles
larger than 4GB. Use the NVMAP_ELEM_SIZE_U64 flag to
differentiate between 32 bit and 64 bit sizes, offsets.

bug 1860962

Reviewed-on: http://git-master/r/1287378
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/video/tegra/nvmap/nvmap_cache.c
@@ -527,8 +527,8 @@ out:
  *
  * NOTE: this omits outer cache operations which is fine for ARM64
  */
-int nvmap_do_cache_maint_list(struct nvmap_handle **handles, u32 *offsets,
-			      u32 *sizes, int op, int nr)
+int nvmap_do_cache_maint_list(struct nvmap_handle **handles, u64 *offsets,
+			      u64 *sizes, int op, int nr)
 {
 	int i;
 	u64 total = 0;
@@ -585,8 +585,8 @@ int nvmap_do_cache_maint_list(struct nvmap_handle **handles, u32 *offsets,
 					nvmap_stats_read(NS_CFLUSH_DONE));
 	} else {
 		for (i = 0; i < nr; i++) {
-			u32 size = sizes[i] ? sizes[i] : handles[i]->size;
-			u32 offset = sizes[i] ? offsets[i] : 0;
+			u64 size = sizes[i] ? sizes[i] : handles[i]->size;
+			u64 offset = sizes[i] ? offsets[i] : 0;
 			int err = __nvmap_do_cache_maint(handles[i]->owner,
 							 handles[i], offset,
 							 offset + size,
