+++ drivers/video/tegra/nvmap/nvmap_priv.h
@@ -491,8 +491,8 @@ void nvmap_flush_cache(struct page **pages, int numpages);
 int nvmap_cache_maint_phys_range(unsigned int op, phys_addr_t pstart,
 		phys_addr_t pend, int inner, int outer);
 
-int nvmap_do_cache_maint_list(struct nvmap_handle **handles, u32 *offsets,
-			      u32 *sizes, int op, int nr);
+int nvmap_do_cache_maint_list(struct nvmap_handle **handles, u64 *offsets,
+			      u64 *sizes, int op, int nr);
 int __nvmap_cache_maint(struct nvmap_client *client,
 			       struct nvmap_cache_op *op);
 int nvmap_cache_debugfs_init(struct dentry *nvmap_root);
@@ -623,12 +623,12 @@ static inline struct page **nvmap_pages(struct page **pg_pages, u32 nr_pages)
 	return pages;
 }
 
-void nvmap_zap_handle(struct nvmap_handle *handle, u32 offset, u32 size);
+void nvmap_zap_handle(struct nvmap_handle *handle, u64 offset, u64 size);
 
 void nvmap_vma_open(struct vm_area_struct *vma);
 
-int nvmap_reserve_pages(struct nvmap_handle **handles, u32 *offsets,
-			u32 *sizes, u32 nr, u32 op);
+int nvmap_reserve_pages(struct nvmap_handle **handles, u64 *offsets,
+			u64 *sizes, u32 nr, u32 op);
 
 static inline void nvmap_kmaps_inc(struct nvmap_handle *h)
 {
